worker/requirements.txt
pydub
scipy
numpypy
loudnorm
google-cloud-storage
Flask==2.3.2
gunicorn
tqdm